Upgrade ConditionVariable ID and Condition parameter to uint16_t to fix RCN-227 overflows.
- Updated `Condition` struct to use `uint16_t` for `parameter`.
- Updated `ConditionVariable` struct to use `uint16_t` for `id`.
- Updated `MappingRule` to use `std::vector<uint16_t>` for conditions.
- Updated `AuxController` methods and `_cv_states` map to use `uint16_t`.
- Fixed potential overflows in `parseRcn227PerOutputV3` and `V1` by removing truncating casts or using wider types.
- Verified with existing tests in `test_rcn227.cpp`.
- Updated `docs/IMPLEMENTATION_STATUS.MD` to reflect Phase 3 and 4 completion.

+++ firmware/lib/xDuinoRails_DccLightsAndFunctions/src/AuxController.h
@@ -100,7 +100,7 @@ public:
      * @param cv_id The ID of the ConditionVariable.
      * @return The cached boolean result of the variable's evaluation.
      */
-    bool getConditionVariableState(uint8_t cv_id) const;
+    bool getConditionVariableState(uint16_t cv_id) const;
     /**
      * @brief Gets the current value of a binary state.
      * @param state_number The ID of the binary state to check.
@@ -144,7 +144,7 @@ private:
     DecoderDirection _direction = DECODER_DIRECTION_FORWARD;
     uint16_t _speed = 0;
     std::map<uint16_t, bool> m_binary_states;
-    std::map<uint8_t, bool> _cv_states;
+    std::map<uint16_t, bool> _cv_states;
     bool _state_changed = true;
 };
